Multilingual Directions WepApp/requirements.txt
flask
gunicorn
google-cloud-translate
